bug fix: server error while empty articles on home page

+++ helloscala-backend/server/src/main/java/com/helloscala/common/mapper/ArticleMapper.java
@@ -29,7 +29,7 @@ public interface ArticleMapper extends BaseMapper<Article> {
                                                        @Param("orderByDescColumn")String orderByDescColumn);
 
 
-    ArticleInfoVO selectArticleByIdToVO(Integer id);
+    ArticleInfoVO selectArticleByIdToVO(Long id);
 
     List<ArticleVO> selectListByBanner();
